REFACTOR playRound as eventhandler on button press
The playRound button had previously been called during runtime within
the game function. The code is now initiated by the player selecting a
hand to play. Therefore I decided to refactor the playRound function as
an eventhandler for the buttons. The console currently outputs the
winner of the round and the current score.